Fix warnings and upgrade to C++20
[[unlikely]] requires c++20

+++ src/logic/pakfile.h
@@ -106,7 +106,7 @@ public:
 class CPakFile
 {
 public:
-	CPakFile(int version)
+	CPakFile(short version)
 	{
 		m_Header.fileVersion = version;
 	};
